# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SpaceX Pre-IPO Share Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List user's purchase requests
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseRequest"
        "401":
          description: Unauthorized
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Submit a share purchase request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Purchase request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  # --- Dashboard ---
...
components:
  schemas:
...
    PurchaseRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amountUsd:
          type: number
        requestedShares:
          type: number
        pricePerShare:
          type: number
        status:
          type: string
          enum: [pending_review, confirmed, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amountUsd
        - requestedShares
        - pricePerShare
        - status
        - createdAt
    PurchaseInput:
      type: object
      properties:
        amountUsd:
          type: number
          minimum: 5000
        agreedToTerms:
          type: boolean
      required:
        - amountUsd
        - agreedToTerms
...