# lib/api-zod/src/generated/types/purchaseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SpaceX Pre-IPO Share Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseInput {
  /** @minimum 5000 */
  amountUsd: number;
  agreedToTerms: boolean;
}