# lib/api-zod/src/generated/types/purchaseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SpaceX Pre-IPO Share Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseRequestStatus } from "./purchaseRequestStatus";

export interface PurchaseRequest {
  id: number;
  userId: number;
  amountUsd: number;
  requestedShares: number;
  pricePerShare: number;
  status: PurchaseRequestStatus;
  createdAt: Date;
}